Implemented playBtn logic and mainGame load
playBtn now loads the loading scene, which in turn loads the mainGame.
Loading scene is there in case in mainGame takes a while to load, we
need to not disorient the player. This works by transferring the
OVRCameraRig through all three levels, then disposing of it once
mainGame is loaded. mainGame then switches to its own camera setup and
unFades from black to play the game.